Implement user prefs and browser notifications
Add a new user preferences management system to the app :
- Add a new "My options" section that is shown only in the user page of the
current user (aka "My profile" link in the header).
- Add two new api routes /api/preferences/set/<key>/<value> and
/api/preferences/get/<key>, respectively used to fetch and set preferences
values. Prefs are also present in jsvars values present in the pages.
- These prefs are stored in the user session (aka data stored server-side
identified by an id stored locally in a cookie, like where is currently
stored the username after being logged-in). Note that this storage is
therefore not permanently linked to the user, and is currently completely
deleted at logout via the "logout" menu.
- Create a "user.js" file that is loaded when user is watching user pages,
used to manage options checkboxes.

Add a new pref corresponding to the notifications :
- Create a "useNotifications" option that if activated correspond to the
presence of a not null value of the "useNotifications" pref.
- Add the usage of status notifications in query view page, when the query is
completed and if the option was enabled previously by the user (presence of
"useNotifications" pref).
- To avoid triggering it if the query was already ended, checkStatus fct has
now a "silent" param.

Bug: T124625
Change-Id: Iaf44023f67144c17924275b799ad687a2e357d4a