Style changes regarding navbar and maintenance-msg
The maintenance-msg is very large and covers some parts of the page
that was padded by 96px. The default font size is small enough to
not too much of the area. Having bold increases the significance.

Additionally, the maintenance-msg is made to not stick, having
position as default static. The navbar is set to have position
sticky. Both bahaviors reserve an area in the body, unlike position
absolute, and no padding in #content is required. However, position
sticky is not supported in IE, and in such case it defaults to
static.

Change-Id: I4711660295a7fe71c809e8aeb2d9d7455f1ae583