results: Fixes for memory consumption and bytes support
The old implementations for various file types put all the results
into an in-memory buffer. This can get very troublesome for larger
result sets. The new implementation will make the responses all
generators as far as flask and xlsxwriter see. Some implementations
use a file-like object for writing and to make it have a generator
interface werkzeug's IterIO is used, and it supports both bytes
and string (unicode), but requires the installation of greenlet,
which has been added to Pipfile and requirements.txt. A subclass
_IterI is added so that it flushes on every 1024 pos added.

JSON serialization now uses a custom 'fake list' implementation to
serialize generators. It implements __iter__ and __bool__, and is a
(bad) subclass of list, so and json will just pretend it's a list
and be good with it without loading the whole list into memory.
Bytes as part of the result set also used to cause issues in JSON
serialization because it's not technically serializable in Python
3. We now convert it to string (unicode) with utf-8 decoding, and
surrogateescape anything that cannot be decoded by utf-8. A helper
function _stringify_results is added to handle bytes in non-JSON
cases.

Bug: T192696
Change-Id: I8acb929d78c7d62815c5ca38f1f1e68bd2a4b0b3