Handle bad output_result endpoint params
An empty query return no resultset.
This patch handle the "no such table: resultset_0" sqlite error
that appear when output of such query is loaded, by returning
a 404 error.
Note that it's not possible to extract better/properly sqlite
error codes, see https://stackoverflow.com/a/38847510/8966159.
The client now avoids doing an unnecessary request if there is
nothing to fetch, and shows a "This query returned no results."
message. And in any case the 404 is gracefully handled
by the client (view.js) and will not show any resultset.

At the same time invalid query_run id case is handled
client-side.

Bug: T205222
Change-Id: Id9a4da1329237c44965bdbeb836ac5e06478eb75