multiinstance: Attempt to make quarry work with multiinstance replicas
A pending redesign of the wiki replicas will require connections to
the section (or wiki database name) the user intends to connect to in
order to connect at all. This requires a fundamental change in how
quarry works. This also means the existing killer.py simply won't work, and
quarry may need to just rely on the pt-kill settings we have on the replicas
themselves these days.

To deploy this, the query_revision table must be altered to include the
new query_database column. The config.yaml needs to include a REPLICA_DOMAIN
value that currently would match ".db.svc.eqiad.wmflabs". A new local
vagrant instance should Just Work as long as you specify the database is
"quarry".

When finally deployed, this will require users to refresh the page if they
have loaded it previously to get the changes to view.js.

I don't think this needs to be merged before the replicas are deployed as
multiinstance, but it could be! Any wiki database could be entered and
should work, since you'll be connecting to all of them for the time being.

WIP: An untested multiinstance_migration.py script is in this patch. Will be testing
before merge, but feel free to offer thoughts on what it does so far.

For more context: https://wikitech.wikimedia.org/wiki/News/Wikireplicas_2020_Redesign

Bug: T264254
Change-Id: Ia00cb82ffe4ebb920d10660d562035f673115d33