Add database autocompletion
This is done by:
* Adding a new api endpoint to get the list of databases
* Adding the autocomplete functionality to the database selector input

The databases are fetch from the known databases from query_revisions,
so it really helps creating the following index:
```
CREATE INDEX IF NOT EXISTS query_rev_query_database_index ON query_revision(query_database);
```

The fetching of the databases is asynchronous (ajax) not affecting the
page load time.

NOTE: this does not validate that the databases are correct, as the
data in the query_revisions table might not be either.

Screenshot: F34591751

This includes some fixes that the previous attempt did not have:
* Show correctly the db name when loading an existing query.

Bug: T287471
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: I3c6f9ab122354decfb05e677133b263c53bfd79f