celery update
celery is working in the container, though does not start
due to a syntax error in the vm. This may be due to python
3.7 introducing "async" as a keyword, which redis.py had
been using for itself. It remains unclear on why this does
not fail on the container.

Would appear that an ordering of redis variables escaped
detection in the containers. Updated here.

Change-Id: Ie6b60295ebab81806b28d7edf1135d51f2f482b3