local dev: set up a really basic wikireplica-ish container
This changes the way the dev environment works to be more similar to prod.

Instead of querying its own database (with ALL grants) in a confusing way,
this uses a modified dump of a cheesy local mediawiki instance with some
sample data. The local user has read-only privileges on this database.

This also adds some documentation for using it.

Bug: T287902
Change-Id: Ie2119985bc8e0b0f8ab1fc5f45a7eacd80d29dc0