close quarry db dropdown on tab
DB selection dropdown list does not close on a tab press.
This change will close when tab is pressed. If nothing from
the dropdown is selected, the dropdown is simply closed, and
the tab acts as normal moving to the next item on the
screen. If something was selected with the arrow keys a tab
will autofill the text box before moving focus to the next
item.

Bug: T289872
Change-Id: Ib07ee362ab279cc5ddc86bed318833248fd607f9