Use vars.qrun_id when stopping query
While looking in to T307297 I noticed requests to
http://localhost:5000/run/undefined/status and noticed that the data
coming back from the /api/query/stop request looks like

    {"stopped": "job stopped"}

It does not have a `qrun_id` field.

`qrun_id` is available in the higher scope under `vars.qrun_id`; let's
use that.

I'm not sure if the query appears to not be stopped because of this bug
or if there's another issue, but this at least fixes the 404 errors that
query for `run/undefined/status`.

Bug: T307297
Change-Id: I459db287b1b9f42996f70cdaa4f19661a731217d